qc: relocate 'fastqc' module to 'qc.apps'.
